requirements.txt
azure-ai-documentintelligence==1.0.2
azure-cognitiveservices-speech==1.48.2
azure-core==1.38.2
azure-search-documents==11.6.0
Flask==3.0.1
Flask-Cors==4.0.0
gunicorn==25.1.0
langchain==1.2.10
langchain-anthropic==0.3.16
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-google-genai==4.2.1
langchain-huggingface==0.3.0
langchain-openai==0.1.7
langchain-text-splitters==1.1.1
pypdf==6.7.5
python-dotenv==1.0.1
